+++ lib/common/common/src/counter/mod.rs
@@ -2,4 +2,5 @@ pub mod counter_cell;
 pub mod hardware_accumulator;
 pub mod hardware_counter;
 pub mod hardware_data;
+pub mod iterator_hw_measurement;
 pub mod referenced_counter;
