+++ lib/segment/src/index/field_index/map_index/mmap_map_index.rs
@@ -6,7 +6,7 @@ use std::path::{Path, PathBuf};
 
 use ahash::HashMap;
 use common::counter::hardware_counter::HardwareCounterCell;
-use common::iterator_ext::IteratorExt;
+use common::counter::iterator_hw_measurement::HwMeasurementIteratorExt;
 use common::mmap_hashmap::{Key, MmapHashMap, READ_ENTRY_OVERHEAD};
 use common::types::PointOffsetType;
 use io::file_operations::{atomic_save_json, read_json};
